Executable version instead of just saving a copy.
